refactor: split campaign interfaces by concern
